Using just one GCPLog target with multiple subs types
