+++ clients/pkg/promtail/targets/gcplog/push_translation.go
@@ -1,4 +1,4 @@
-package gcppush
+package gcplog
 
 import (
 	"encoding/base64"
