+++ clients/pkg/promtail/targets/gcplog/push_translation_test.go
@@ -1,4 +1,4 @@
-package gcppush
+package gcplog
 
 import (
 	"testing"
